backend/requirements.txt
supabase==2.3.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
python-amazon-paapi==5.0.1 